requirements.txt
dvc
tensorflow
numpy
pandas
matplotlib
tqdm
PyYAML
boto3

-e .